fix: move Number to fun.
